+++ src/helpers/executeCommand.mjs
